I commit at 17:51 /add render contract

+++ App/src/main/java/khoindn/swp391/be/app/model/Response/ContractHistoryRes.java
@@ -13,7 +13,7 @@ import java.time.LocalDate;
 public class ContractHistoryRes {
     private int contractId;
     private String vehicleName;
-    private float ownership;
+    private int ownership;
     private String status;
     private LocalDate signedAt;
 }
